Start a ripping leftwm-core
