--- leftwm-core/src/config/keybind.rs
@@ -1,9 +1,0 @@
-use crate::Command;
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct Keybind {
-    pub command: Command,
-    pub modifier: Vec<String>,
-    pub key: String,
-}
