+++ leftwm-core/src/config/mod.rs
@@ -1,4 +1,3 @@
-mod keybind;
 mod scratchpad;
 mod workspace_config;
 
@@ -7,14 +6,10 @@ use crate::layouts::Layout;
 pub use crate::models::{FocusBehaviour, Gutter, Margins, Size};
 use crate::models::{LayoutMode, Manager, Window, WindowType};
 use crate::state::State;
-pub use keybind::Keybind;
 pub use scratchpad::ScratchPad;
 pub use workspace_config::Workspace;
 
 pub trait Config {
-    /// Returns a collection of bindings with the mod key mapped.
-    fn mapped_bindings(&self) -> Vec<Keybind>;
-
     fn create_list_of_tag_labels(&self) -> Vec<String>;
 
     fn workspaces(&self) -> Option<Vec<Workspace>>;
