+++ leftwm-core/src/display_action.rs
@@ -1,4 +1,3 @@
-use crate::config::Keybind;
 use crate::models::TagId;
 use crate::models::Window;
 use crate::models::WindowHandle;
@@ -64,9 +63,6 @@ pub enum DisplayAction {
     /// Tell the DM to return to normal mode if it is not (ie resize a
     /// window or moving a window).
     NormalMode,
-
-    /// SoftReload keygrabs, needed when keyboard changes.
-    ReloadKeyGrabs(Vec<Keybind>),
 
     /// Configure a xlib window.
     ConfigureXlibWindow(Window),
