+++ leftwm-core/src/utils/mod.rs
@@ -2,6 +2,6 @@
 pub mod child_process;
 pub mod command_pipe;
 pub mod helpers;
+pub mod modmask_lookup;
 pub mod state_socket;
 pub mod window_updater;
-pub mod xkeysym_lookup;
